Refactored data sorting:  - Moved from Summaries Controller to new service class  - Updated specs to match
